requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
idna==3.11
msal==1.35.1
o365==2.1.9
pycparser==3.0
PyJWT==2.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
requests==2.32.5
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
